refactor(lazy-tags): remove public impl header
